requirements.txt
aiofiles==0.6.0
alembic==1.4.3
click==7.1.2
fastapi==0.61.2
h11==0.11.0
Mako==1.1.3
MarkupSafe==1.1.1
mysql-connector-python==8.0.22
Pillow==8.0.1
protobuf==3.13.0
pydantic==1.7.2
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
python-multipart==0.0.5
six==1.15.0
SQLAlchemy==1.3.20
starlette==0.13.6
uvicorn==0.12.2
requests==2.25.0
